#90 change components structure
