--- src/views/CategoryView.vue
@@ -1,7 +1,0 @@
-<template>
-  <QuestionsList />
-</template>
-
-<script setup>
-import QuestionsList from "../components/questions/QuestionsList.vue";
-</script>
