--- src/views/HomeView.vue
@@ -1,3 +1,0 @@
-<template>
-  <h1>Home</h1>
-</template>
