--- src/views/QuizView.vue
@@ -1,9 +1,0 @@
-<template>
-  <div>
-    <h1>Quiz</h1>
-  </div>
-</template>
-
-<script setup></script>
-
-<style></style>
